+++ anaconda_linter/lint/check_syntax.py
@@ -25,12 +25,14 @@ class version_constraints_missing_whitespace(LintCheck):
         for section in ("build", "run", "host"):
             check_paths.append(f"requirements/{section}")
 
-        constraints = re.compile("(.*?)([<=>].*)")
+        constraints = re.compile("(.*?)([!<=>].*)")
         for path in check_paths:
             for n, spec in enumerate(recipe.get(path, [])):
                 has_constraints = constraints.search(spec)
                 if has_constraints:
-                    space_separated = has_constraints[1].endswith(" ")
+                    # The second condition is a fallback.
+                    # See: https://github.com/anaconda-distribution/anaconda-linter/issues/113
+                    space_separated = has_constraints[1].endswith(" ") or " " in has_constraints[0]
                     if not space_separated:
                         self.message(section=f"{path}/{n}", data=True)
 
@@ -39,12 +41,14 @@ class version_constraints_missing_whitespace(LintCheck):
         for section in ("build", "run", "host"):
             check_paths.append(f"requirements/{section}")
 
-        constraints = re.compile("(.*?)([<=>].*)")
+        constraints = re.compile("(.*?)([!<=>].*)")
         for path in check_paths:
             for spec in self.recipe.get(path, []):
                 has_constraints = constraints.search(spec)
                 if has_constraints:
-                    space_separated = has_constraints[1].endswith(" ")
+                    # The second condition is a fallback.
+                    # See: https://github.com/anaconda-distribution/anaconda-linter/issues/113
+                    space_separated = has_constraints[1].endswith(" ") or " " in has_constraints[0]
                     if not space_separated:
                         dep, ver = has_constraints.groups()
                         self.recipe.replace(spec, f"{dep} {ver}", within="requirements")