+++ anaconda_linter/lint/check_url.py
@@ -5,7 +5,7 @@ Verify that the URLs in the recipe are valid
 """
 
 from .. import utils
-from . import WARNING, LintCheck
+from . import ERROR, INFO, WARNING, LintCheck
 
 
 class invalid_url(LintCheck):
@@ -21,6 +21,7 @@ class invalid_url(LintCheck):
             response_data = utils.check_url(url)
             acceptable_redirects = [
                 ("pypi.io", "pypi.org"),
+                ("pypi.org", "files.pythonhosted.org"),
                 ("github.com", "objects.githubusercontent.com"),
                 ("github.com", "codeload.github.com"),
             ]
@@ -35,7 +36,8 @@ class invalid_url(LintCheck):
                 self.__class__.__doc__ = self.__class__.__doc__.format(
                     url, response_data["message"]
                 )
-                self.message(section=section)
+                severity = INFO if "domain_redirect" in response_data else ERROR
+                self.message(section=section, severity=severity)
 
     def check_recipe(self, recipe):
         url_fields = [
@@ -50,10 +52,13 @@ class invalid_url(LintCheck):
             if url:
                 response_data = utils.check_url(url)
                 if response_data["code"] < 0 or response_data["code"] >= 400:
+                    reset_text = self.__class__.__doc__
                     self.__class__.__doc__ = self.__class__.__doc__.format(
                         url, response_data["message"]
                     )
-                    self.message(section=url_field)
+                    severity = INFO if "domain_redirect" in response_data else ERROR
+                    self.message(section=url_field, severity=severity)
+                    self.__class__.__doc__ = reset_text
 
 
 class http_url(LintCheck):
@@ -62,8 +67,6 @@ class http_url(LintCheck):
     Please replace with https if possible.
 
     """
-
-    severity = WARNING
 
     def check_source(self, source, section):
         url = source.get("url", "")
@@ -82,5 +85,7 @@ class http_url(LintCheck):
         for url_field in url_fields:
             url = recipe.get(url_field, "")
             if url.lower().startswith("http://"):
+                reset_text = self.__class__.__doc__
                 self.__class__.__doc__ = self.__class__.__doc__.format(url)
-                self.message(section=url_field.split("/")[0])
+                self.message(section=url_field.split("/")[0], severity=WARNING)
+                self.__class__.__doc__ = reset_text