+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 Note: version releases in the 0.x.y range may introduce breaking changes.
 
+## 0.0.4
+
+- Add test suite
+- Update documentation
+- Bug fixes for message system
+- Add new check: `patch_unnecessary`
+- Add `--severity` option to control the severity level of linter messages
+- Remove `load_skips` function
+
 ## 0.0.3
 
 - Update installation files